Fixed status change bug

+++ workspaces/server/models/requests.go
@@ -5,8 +5,8 @@ import "gorm.io/gorm"
 type CustomerRequest struct {
 	gorm.Model
 	RequestID   string `gorm:"primary_key" json:"requestID"`
-	EmployeeID  int64  `json:"employeeID"`
-	RoomID      int64  `json:"roomNumber"`
+	EmployeeID  int    `json:"employeeID"`
+	RoomID      int    `json:"roomNumber"`
 	Email       string `json:"email"`
 	RequestType string `json:"requestType"`
 	Comment     string `json:"comment"`
